+++ src/nclutils/text/__init__.py
@@ -1,5 +1,5 @@
 """Document utilities."""
 
-from .text_processing import ensure_lines_in_file, replace_in_file
+from .text import ensure_lines_in_file, replace_in_file
 
 __all__ = ["ensure_lines_in_file", "replace_in_file"]
